libvdsk/qcow: Move qcow specific structures to own header file
